fix: eliminate all Java compiler warnings and Gradle DSL deprecations
Agent-Logs-Url: https://github.com/drake7707/MidiSheetMusic-Android/sessions/305dc9a5-59f4-416f-8f7c-42ed2e8da8f7

+++ app/src/main/java/com/midisheetmusic/MidiFile.java
@@ -642,7 +642,7 @@ public class MidiFile {
                           " != 4", file.GetOffset());
                     }
                     else {
-                        mevent.Numerator = ((byte)mevent.Value[0]);
+                        mevent.Numerator = mevent.Value[0];
                         mevent.Denominator = ((byte)Math.pow(2, mevent.Value[1]));
                     }
                 }
